rename(#100): ProductSearchDto를 utils 패키지로 이동
