+++ service/product/server/src/main/java/com/sparta/product/domain/repository/ElasticSearchRepository.java
@@ -1,6 +1,6 @@
 package com.sparta.product.domain.repository;
 
-import com.sparta.product.infrastructure.elasticsearch.dto.ProductSearchDto;
+import com.sparta.product.infrastructure.utils.ProductSearchDto;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface ElasticSearchRepository
